Fix OpenCvSharp4 and ImGui.NET API compatibility errors
Resolved all compilation errors from API changes in OpenCvSharp4 and ImGui.NET.

## OpenCV API Fixes

### 1. DepthAwareRansac.cs
**Problem**: FindEssentialMat and RecoverPose had outdated signatures
- Old API used `FundamentalMatMethods` enum and direct array passing
- New API requires `EssentialMatMethod` (singular) and InputArray wrappers

**Solution**:
```csharp
// Before:
var E = Cv2.FindEssentialMat(points1, points2, cameraMatrix,
    FundamentalMatMethods.Ransac, confidence, threshold, out var mask);

// After:
var mask = new Mat();
var E = Cv2.FindEssentialMat(
    InputArray.Create(points1),
    InputArray.Create(points2),
    InputArray.Create(cameraMatrix),
    EssentialMatMethod.Ransac,
    confidence, threshold, mask);
```

RecoverPose also updated with InputArray/OutputArray wrappers.

### 2. MatTextureConverter.cs
**Problem**: `mat.Step()` returns `long` but was assigned to `int`

**Solution**: Added explicit cast `(int)` in 3 locations:
- Line 90: ConvertGrayscaleToRGBA
- Line 119: ConvertBGRToRGBA
- Line 147: ConvertBGRAToRGBA

```csharp
int srcIdx = (int)(y * mat.Step() + x);
```

### 3. KeyframeManager.cs
**Problem**: `List<Point3f>` not assignable to InputArray parameter in SolvePnPRansac

**Solution**: Convert lists to arrays and wrap with InputArray.Create():
```csharp
Cv2.SolvePnPRansac(
    InputArray.Create(objectPoints.ToArray()),
    InputArray.Create(imagePoints.ToArray()),
    cameraMatrix, ...);
```

### 4. GeoreferencingManager.cs
**Problem**: `Quaternion.TryCreateFromRotationMatrix` doesn't exist in .NET

**Solution**: Use `CreateFromRotationMatrix` with try-catch:
```csharp
try
{
    rotation = Quaternion.CreateFromRotationMatrix(rotMat);
}
catch
{
    rotation = Quaternion.Identity;
}
```

## ImGui API Fixes

### 1. RadioButton Type Mismatch
**Problem**: RadioButton requires `ref int`, not `ref bool`

**Solution**: Changed `_useVideoFile` from `bool` to `int`:
```csharp
// Declaration
private int _useVideoFile; // 0 = webcam, 1 = file

// Usage
ImGui.RadioButton("Webcam/Camera", ref _useVideoFile, 0);
ImGui.RadioButton("Video File", ref _useVideoFile, 1);

// Comparisons
if (_useVideoFile == 0) // webcam
if (_useVideoFile == 1) // file
```

### 2. BeginChild Signature Change
**Problem**: BeginChild third parameter changed from `bool` to `ImGuiChildFlags`

**Solution**:
```csharp
// Before:
ImGui.BeginChild(label, size, true);

// After:
ImGui.BeginChild(label, size, ImGuiChildFlags.Border);
```

## Files Modified

1. **Analysis/Photogrammetry/DepthAwareRansac.cs**
   - Updated FindEssentialMat and RecoverPose calls
   - Added InputArray/OutputArray wrappers

2. **Analysis/Photogrammetry/MatTextureConverter.cs**
   - Added (int) cast for mat.Step() in 3 locations

3. **Analysis/Photogrammetry/KeyframeManager.cs**
   - Converted List to array and wrapped with InputArray.Create

4. **Analysis/Photogrammetry/GeoreferencingManager.cs**
   - Replaced TryCreateFromRotationMatrix with try-catch pattern

5. **UI/Windows/RealtimePhotogrammetryWindow.cs**
   - Changed _useVideoFile from bool to int
   - Updated RadioButton calls and comparisons
   - Fixed BeginChild to use ImGuiChildFlags.Border

## Technical Details

**OpenCvSharp4 Changes**:
- Mat.Step() returns `long` (was `int` in older versions)
- FindEssentialMat/RecoverPose require InputArray wrappers
- EssentialMatMethod enum replaces FundamentalMatMethods for essential matrix

**ImGui.NET Changes**:
- RadioButton<T> removed, now only RadioButton(string, ref int, int)
- BeginChild uses ImGuiChildFlags enum instead of bool for borders

**Compatibility**: These changes maintain backward compatibility with the
existing logic while adapting to the new API requirements.

## Testing Status

All compilation errors resolved. Code should now build successfully with:
- OpenCvSharp4 v4.10.0
- ImGui.NET v1.90.8.1
- .NET 8.0

+++ Analysis/Photogrammetry/MatTextureConverter.cs
@@ -87,7 +87,7 @@ public static class MatTextureConverter
             {
                 for (int x = 0; x < width; x++)
                 {
-                    int srcIdx = y * mat.Step() + x;
+                    int srcIdx = (int)(y * mat.Step() + x);
                     int dstIdx = (y * width + x) * 4;
 
                     byte gray = src[srcIdx];
@@ -116,7 +116,7 @@ public static class MatTextureConverter
             {
                 for (int x = 0; x < width; x++)
                 {
-                    int srcIdx = y * mat.Step() + x * 3;
+                    int srcIdx = (int)(y * mat.Step() + x * 3);
                     int dstIdx = (y * width + x) * 4;
 
                     pixelData[dstIdx + 0] = src[srcIdx + 2]; // R (from B)
@@ -144,7 +144,7 @@ public static class MatTextureConverter
             {
                 for (int x = 0; x < width; x++)
                 {
-                    int srcIdx = y * mat.Step() + x * 4;
+                    int srcIdx = (int)(y * mat.Step() + x * 4);
                     int dstIdx = (y * width + x) * 4;
 
                     pixelData[dstIdx + 0] = src[srcIdx + 2]; // R (from B)
